# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Workforce API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: rooms
    description: Room operations
  - name: tasks
    description: Task operations
  - name: assignments
    description: Assignment operations
  - name: shifts
    description: Shift operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      description: Authenticate with email and password, returns a bearer token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session
      description: Returns the current user, their memberships, roles, and permissions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/auth/me/access-context:
    get:
      operationId: getMeAccessContext
      tags: [auth]
      summary: Get employment scopes for the current user
      description: |
        Returns the employment scope for the authenticated user by resolving
        user_employee_links → employee_profiles → employee_role_assignments.
        This endpoint separates the system user identity (from /auth/me) from
        employee profiles, preserving the user/employee boundary.

        **Compatibility mode:** When no employee-profile links are found in the
        local database, the endpoint synthesises a single `COMPAT` scope using
        the permissions and roles already carried by the caller's auth token or
        session. This ensures `has_access` is `true` and `active_scope_count` is
        `1` for any user with an active membership, so the frontend
        `employmentScope` is never silently null. No RBAC is bypassed — the
        synthesised scope reflects exactly the permissions present in the token.
        Real employee-profile links (when present) always take precedence.

        **Upstream support:** The local API proxy (api-server) implements this
        endpoint using the local SQLite database. Support on the upstream
        PythonAnywhere backend (hn3t.pythonanywhere.com) is **not yet available**.
        Until the upstream backend adds this endpoint, callers using PythonAnywhere
        JWT tokens are served the compatibility response described above.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: |
            Access context resolved. `has_access` is `true` when at least one
            employee-profile link was found **or** a compatibility scope was
            synthesised from the caller's existing membership/permissions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessContext"
        "401":
          description: Unauthorized — no valid bearer token provided
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"


  /v1/auth/switch-business:
    post:
      operationId: switchBusiness
      tags: [auth]
      summary: Switch active business context
      description: Switch to a different business membership and returns a new token
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchBusinessRequest"
      responses:
        "200":
          description: Business switched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not a member of this business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/rooms/:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List rooms
      description: Returns a list of rooms for the active business
      security:
        - bearerAuth: []
      parameters:
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/tasks/:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      description: Returns a list of tasks for the active business
      security:
        - bearerAuth: []
      parameters:
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/assignments/:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List assignments
      description: Returns a list of shift assignments for the active business
      security:
        - bearerAuth: []
      parameters:
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/shifts/:
    get:
      operationId: listShifts
      tags: [shifts]
      summary: List shifts
      description: Returns a list of shifts for the active business
      security:
        - bearerAuth: []
      parameters:
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: List of shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shift"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
      required:
        - detail

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
      required:
        - access_token
        - token_type

    SwitchBusinessRequest:
      type: object
      properties:
        business_id:
          type: string
      required:
        - business_id

    Membership:
      type: object
      properties:
        business_id:
          type: string
        business_name:
          type: string
        role:
          type: string
      required:
        - business_id
        - business_name
        - role

    SessionInfo:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        memberships:
          type: array
          items:
            $ref: "#/components/schemas/Membership"
        active_business_id:
          type: string
        roles:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
      required:
        - id
        - email
        - memberships

    Room:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        room_number:
          type: string
        status:
          type: string
        room_type:
          type: string
        floor:
          type: string
        building:
          type: string
        location_id:
          type: string
        notes:
          type: string
      required:
        - id
        - name

    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        task_type:
          type: string
        due_date:
          type: string
        assigned_to:
          type: string
        created_at:
          type: string
        location_id:
          type: string
        room_id:
          type: string
      required:
        - id
        - title

    Assignment:
      type: object
      properties:
        id:
          type: string
        shift_id:
          type: string
        employee_id:
          type: string
        employee_name:
          type: string
        status:
          type: string
        role:
          type: string
        created_at:
          type: string
      required:
        - id
        - shift_id
        - employee_id

    Shift:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        start_time:
          type: string
        end_time:
          type: string
        status:
          type: string
        role:
          type: string
        location_id:
          type: string
        capacity:
          type: integer
        filled:
          type: integer
        created_at:
          type: string
      required:
        - id
        - start_time
        - end_time

    EmployeeRoleAssignment:
      type: object
      description: A single role assignment entry for an employee
      properties:
        id:
          type: string
        role_name:
          type: string
        scope_type:
          type: string
          enum: [BUSINESS, LOCATION]
        location_id:
          type: string
          nullable: true
        location_name:
          type: string
          nullable: true
        permissions:
          type: array
          items:
            type: string
      required:
        - id
        - role_name
        - scope_type

    EmployeeScope:
      type: object
      description: |
        A single employment scope linking a system user to an employee profile,
        including the employee's role assignments and effective permissions.

        When `link_status` is `"COMPAT"` the scope was synthesised from the
        caller's auth token because no employee-profile link record was found.
        All other fields are populated from real DB records when available.
      properties:
        employee_profile_id:
          type: string
        link_id:
          type: string
        link_status:
          type: string
        business_id:
          type: string
        employee_name:
          type: string
        job_title:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        employee_code:
          type: string
          nullable: true
        employment_status:
          type: string
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeRoleAssignment"
        effective_permissions:
          type: array
          description: Union of all permissions from all role assignments
          items:
            type: string
        is_super_admin:
          type: boolean
          description: True when effective_permissions includes "*"
      required:
        - employee_profile_id
        - link_id
        - link_status
        - business_id
        - employee_name
        - employment_status
        - assignments
        - effective_permissions
        - is_super_admin

    AccessContext:
      type: object
      description: |
        Employment access context for the authenticated user.
        Returned by GET /api/v1/auth/me/access-context.

        The local API proxy (api-server) provides this using local SQLite records.
        When no employee-profile links are present, a single `COMPAT` scope is
        synthesised from the user's auth token permissions so `has_access` is
        `true` for any user with an active membership.
      properties:
        user_id:
          type: string
        has_access:
          type: boolean
          description: True when at least one active employment scope was found
        active_scope_count:
          type: integer
        scopes:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeScope"
        resolved_at:
          type: string
          format: date-time
      required:
        - user_id
        - has_access
        - active_scope_count
        - scopes
        - resolved_at